Apply option property shorthands to Checklist component
